requirements.txt
annotated-types==0.6.0
anyio==4.1.0
certifi==2023.11.17
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.8
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
